+++ app/src/main/java/com/group80/uoftinder/feed/GenerateCompatibilityList.java
@@ -35,7 +35,7 @@ public class GenerateCompatibilityList {
         this.type = this.curUser.getUserType();
         this.usf = new UserScoreFacade(curUser);
         this.userScoreComparator = Comparator.comparing(user -> compScores.get(user));
-//        calculateCompatibilityList();
+        calculateCompatibilityList();
         filteredCompatibilityList = new ArrayList<>();
     }
